requirements.txt
flask==3.0.0
msal==1.26.0
requests==2.31.0
python-dotenv==1.0.0
spacy==3.7.2
pandas==2.1.4
numpy==1.26.2
gunicorn==21.2.0

